# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Collaborative Kanban board application with real-time updates",
        "title": "CollabBoard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/cards/{id}/duplicate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a copy of an existing card in the same list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Duplicate card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card duplicated successfully",
                        "schema": {
                            "$ref": "#/definitions/cards.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid card ID",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/lists/{listId}/cards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all cards in a specific list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Get list cards",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cards",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cards.CardResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new card in a specific list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Create a new card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cards.CreateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Card created successfully",
                        "schema": {
                            "$ref": "#/definitions/cards.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/lists/{listId}/cards/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update card title, description, position, or list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Update card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cards.UpdateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card updated successfully",
                        "schema": {
                            "$ref": "#/definitions/cards.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a card from the list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Delete card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/cards.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/lists/{listId}/cards/{id}/move": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move a card to a new position within the same list or to a different list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Move card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Current List ID",
                        "name": "listId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Move details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cards.MoveCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card moved successfully",
                        "schema": {
                            "$ref": "#/definitions/cards.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/cards.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cards.CardResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "The login form is not validating email properly"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "listId": {
                    "type": "integer",
                    "example": 1
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Fix login bug"
                }
            }
        },
        "cards.CreateCardRequest": {
            "type": "object",
            "required": [
                "position",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "The login form is not validating email properly"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Fix login bug"
                }
            }
        },
        "cards.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Card not found"
                }
            }
        },
        "cards.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "card deleted"
                }
            }
        },
        "cards.MoveCardRequest": {
            "type": "object",
            "required": [
                "position"
            ],
            "properties": {
                "listId": {
                    "type": "integer",
                    "example": 2
                },
                "position": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "cards.UpdateCardRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Updated description of the bug"
                },
                "listId": {
                    "type": "integer",
                    "example": 3
                },
                "position": {
                    "type": "integer",
                    "example": 2
                },
                "title": {
                    "type": "string",
                    "example": "Fix login validation bug"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}